Use original source file; fix SerialPortDlg
The SerialPortDlg was using unitialized settings. Fix that by
adding a set_settings() function which is called from qflexprop.cpp.

Instead of storing a copy of the source in /tmp/QFlexProp/XXXXXXXX.*
compile the original source file and make sure it is not removed.
Load and remove the intermediate and result files.