Fix initial size of terminal (vt220 class)
Fix some missing widget resizing whenever the terminal's font or
geometry changes. The UpdateSize signal needs the widget to already
be resized or it will use the wrong size.

The cursor is still not always visible probably because when the
terminal is resized (by scrolling up and adding a line) the new
position is outside of the current geometry. Perhaps try to delay
the UpadteCursor signal through a QTimer::singleShot() to have
the event queue handled between the two signals.