term: fix cursor updating by queueing UpdateCursor signal
The UpdateCursor signal is emitted only after the current queue
of events becomes empty. This is important for e.g. adding a line
to the terminal which changes it's widget size.
The new cursor rectangle then is outside of the old widget geometry.
By using a QTimer::singleShot(0, ...) the event queue is emptied
before the UpdateCUrsor signal is sent to the vtScrollArea.